Fix PostgreSQL compatibility for Telegram IDs and Alembic revisions

+++ app/models/user.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import DateTime, Enum, Integer, Numeric, String
+from sqlalchemy import BIGINT, DateTime, Enum, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -12,7 +12,7 @@ class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True, index=True, nullable=False)
+    telegram_id: Mapped[int] = mapped_column(BIGINT, unique=True, index=True, nullable=False)
     username: Mapped[str | None] = mapped_column(String(255), nullable=True)
     language: Mapped[Language] = mapped_column(Enum(Language), default=Language.RU, nullable=False)
     currency: Mapped[Currency] = mapped_column(Enum(Currency), default=Currency.RUB, nullable=False)
